feat(simulator): scope all things
Starting with this change all activities simulated are scopes rather
than plain tokens. This heavily simplifies the mental model of the
catual simulation and better aligns with the BPMN 2.0 specification and
the behavior described therein.

As a positive side-effect this ensures that scopes can be completed
recursively, if finished, fixing a number of quirks with event
sub-processes.

Improvements shipped:

* Event sub-processes work as specified in the BPMN 2.0 standard
* All activities span scopes
* Scope completion invokes exit behaviors recursively
* To start a simulation the started process is now signaled, not
  the contained start event.
* Extract event behaviors into shared component
* Extract scope enter and exit behavior into component