feat(simulator): support compensation
This builds generic compensation support on top of the existing 
transaction support.

Compensation and transaction are related but not the same: While 
mechanisms to register compensate activities are the same, transactions
give special guarantees: At the end of the transaction, it is either
successfully completed (no compensation), canceled (compensating all
successfully completed activites) or errored (no compensation taken 
place).

Compensation is different in a sense that compensation can happen 
as long as the parent scope is not dead. Given our existing execution
model this implies that scopes with compensation activities attached 
do not get destroyed but marked as _compensable_, hence turning 
inactive.

Kept alive as inactive they can be compensated at a future point in
time. They will be destroyed at last once compensation completes
or the parent gets destroyed.

One similarity between compensation and transaction canceling exists:
Both execute compensate activities synchronously and in reverse order
of their registration. To implement this behavior we needed to extend
the existing throw event semantics to await the throw completion, if 
need be.

Related to https://github.com/bpmn-io/bpmn-js-token-simulation/issues/76