feat(simulator): support transactions + cancellation events
This implements basic support for transaction and cancellation events.
More work must be put into the topic to fully implement all cancelation
and compensation scenarios.

In order to make the basic implementation work we had to extend our
existing scope subscription mechanism: It now allows to hook in pre-exit
and a dedicated transaction behavior uses this mechanism to scaffold a
compensation sequence, if the transaction was canceled.

Compensation is done once the process exited (waiting at the cancelation
boundary). Thus, I've built a mechanism to await actual exit on top of
the existing subscription logic, too.

Related to https://github.com/bpmn-io/bpmn-js-token-simulation/issues/76