fix(animation): don't trigger <done> when clearing animation
Previously we did not properly separate completion from clearing an
animation; we would not properly be able to distinguish both,
effectively always triggering the animation <done> effect, also when
clearing a scope (i.e. because it was terminated).

This change moves the control over individual animations to the
animation service, and applies <done> only when the animation completes
normally.